+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 7/27/2022 v 1.0.5
+
+- Fixed an issue causing sets containing only a single end brace to not be properly escaped
+
 ## 6/23/2022 v 1.0.3
 
 - Fixed an issue causing verbose functionality to incorretly remove escaped backslashes.